Extract BigInt dependent code into a module with ts-nocheck
# Conflicts:
#	packages/devextreme/js/__internal/core/license/rsa_pkcs1_sha1.ts
