+++ packages/devextreme/js/__internal/core/license/rsa_comparison.ts
@@ -1,0 +1,32 @@
+// @ts-nocheck
+
+export interface PublicKey {
+  e: number;
+  n: number[];
+}
+
+interface Args {
+  key: PublicKey;
+  signature: number[];
+  actual: number[];
+}
+export function compare(args: Args): boolean {
+  if (typeof(BigInt) === 'undefined') {
+    return true;
+  }
+  const actual = bigIntFromBytes(args.actual);
+
+  const signature = bigIntFromBytes(args.signature);
+  const exponent = BigInt(args.key.e);
+  const modulus = bigIntFromBytes(args.key.n);
+  const expected = (signature ** exponent) % modulus;
+
+  return expected === actual;
+}
+
+const ZERO = BigInt(0);
+const EIGHT = BigInt(8);
+
+function bigIntFromBytes(bytes: number[]): bigint {
+  return bytes.reduce((acc, cur) => (acc << EIGHT) + BigInt(cur) , ZERO);
+}
